+++ client/src/layout/Navbar.jsx
@@ -4,19 +4,15 @@ import Logo from "../components/Logo";
 import { Link, useLocation } from "react-router-dom";
 import { useRecoilValue } from "recoil";
 import { userInfoAtom } from "../stores/userInfo";
-import patientIcon from "../assets/patient.png";
-import femaleDoctor from "../assets/doctor_female.png";
-import maleDoctor from "../assets/doctor_male.png";
-import { CiSquareChevDown } from "react-icons/ci";
+
 import { path } from "../modules/define/path";
+import NavItem from "../components/NavItem";
+
 const NavBar = () => {
   const pathname = useLocation().pathname;
-  const userInfo = useRecoilValue(userInfoAtom);
-  // const menu = ["로그아웃", "마이페이지", "병원페이지"];
-  const [isOpen, setIsOpen] = useState(false);
-  const toggleDropdown = () => {
-    setIsOpen(!isOpen);
-    console.log(isOpen);
+  const [mobileMenu, setMobileMenu] = useState(false);
+  const handleMenu = () => {
+    setMobileMenu(!mobileMenu);
   };
 
   if (pathname === "/" || pathname === "/signUp" || pathname === " ") {
@@ -24,109 +20,68 @@ const NavBar = () => {
   }
   return (
     <NavBarContainer className="navbar-container">
-      <NavLink to={path.home}>
-        <Logo nav="true" />
-      </NavLink>
-      <NavItems>
-        <NavItem>
-          {userInfo.isDoctor ? (
-            <NavLink to="/adminAppointment">예약/진료관리</NavLink>
+      <Wrapper>
+        <NavLink to={path.home}>
+          <Logo nav="true" />
+        </NavLink>
+        <MobileS>
+          {/* 모바일 화면일 떄  */}
+          {mobileMenu === false ? (
+            <MobileButton onClick={handleMenu}>+</MobileButton>
           ) : (
-            <NavLink to="/searchHospital">병원예약</NavLink>
+            <MobileButton onClick={handleMenu}>-</MobileButton>
           )}
-        </NavItem>
-        <NavItem>
-          {userInfo.isDoctor ? (
-            <NavLink to="/adminWaitingList">환자 대기 관리</NavLink>
-          ) : (
-            <NavLink to="/selfdiagnosis">스마트 문진</NavLink>
-          )}
-        </NavItem>
-        <NavItem>
-          {userInfo.isDoctor ? (
-            <NavLink to="/accounting">매출관리</NavLink>
-          ) : (
-            <NavLink to="/pay">진료비 내기</NavLink>
-          )}
-        </NavItem>
-        <NavItem>
-          {userInfo.isDoctor ? (
-            <NavLink to="/teleConsult">실시간 진료 상담</NavLink>
-          ) : (
-            ""
-          )}
-        </NavItem>
-        {/* TODO:여기가 드롭다운이 생길 부분 */}
-        <NavItem>
-          {userInfo.isDoctor ? (
-            userInfo.gender === "female" ? (
-              <NameContainer>
-                <Icon src={femaleDoctor} alt="여자의사" />
-                <div>{userInfo.username}의사</div>
-                <CiSquareChevDown onClick={toggleDropdown} />
-              </NameContainer>
-            ) : (
-              <NameContainer>
-                <Icon src={maleDoctor} alt="남자의사" />
-                <div>{userInfo.username}의사</div>
-                <CiSquareChevDown onClick={toggleDropdown} />
-              </NameContainer>
-            )
-          ) : (
-            <NameContainer>
-              <Icon src={patientIcon} alt="환자아이콘" />
-              <div>{userInfo.username}환자</div>
-              <CiSquareChevDown onClick={toggleDropdown} />
-            </NameContainer>
-          )}
-          {/* TODO:드롭다운이 생길부분 */}{" "}
-          <MenuContent open={isOpen}>
-            <MenuLinkContainer>
-              <SignoutButton
-                onClick={() => {
-                  console.log("logout");
-                }}
-              >
-                로그아웃
-              </SignoutButton>
-              {userInfo.isDoctor ? (
-                <MenuLink to="/hospitalPage">병원페이지</MenuLink>
-              ) : (
-                <MenuLink to="/mypage">마이페이지</MenuLink>
-              )}
-            </MenuLinkContainer>
-          </MenuContent>
-        </NavItem>
-      </NavItems>
+        </MobileS>
+        <Desktop>
+          <NavItem />
+        </Desktop>
+      </Wrapper>
+      <Mobile>
+        {/* 모바일일 때  */}
+        {mobileMenu === false ? null : <NavItem />}
+      </Mobile>
     </NavBarContainer>
   );
 };
 
 export default NavBar;
 const NavBarContainer = styled.nav`
-  background-color: #fff;
-  padding: 1rem;
-  height: 100px;
-  display: flex;
-  justify-content: space-between;
+  position: relative;
+  z-index: 10;
+  width: 100%;
+  color: white;
   box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
 `;
 
-const NavItems = styled.div`
-  list-style: none;
+const Wrapper = styled.div`
   display: flex;
-  gap: 60px;
   align-items: center;
-  margin: 0;
-  padding: 0;
-  flex-direction: row;
+  justify-content: space-between;
+`;
+const MobileS = styled.div`
+  font-size: 1.5rem;
+  display: none;
+  @media (max-width: 640px) {
+    display: block;
+  }
+`;
+const MobileButton = styled.button`
+  background-color: none;
 `;
 
-const NavItem = styled.div`
-  margin-right: 1rem;
-  &:last-child {
-    position: relative;
-    z-index: 2;
+const Desktop = styled.div`
+  display: none;
+
+  @media (min-width: 640px) {
+    display: block;
+  }
+`;
+
+const Mobile = styled.div`
+  display: block;
+
+  @media (min-width: 640px) {
+    display: none;
   }
 `;
 
@@ -135,49 +90,8 @@ const NavLink = styled(Link)`
   text-decoration: none;
   font-weight: 700;
   font-size: 1.5rem;
-  white-space: nowrap; /* 줄 바꿈 방지 */
+  white-space: nowrap;
   @media (max-width: 768px) {
     font-size: 0.5rem;
-  }
-`;
-
-const Icon = styled.img`
-  width: 2rem;
-`;
-const NameContainer = styled.div`
-  display: flex;
-  align-items: center;
-  font-size: 1.5rem;
-  white-space: nowrap;
-  font-weight: 600;
-  &:hover {
-    cursor: pointer;
-  }
-`;
-const MenuContent = styled.div`
-  display: ${(props) => (props.open ? "block" : "none")};
-  position: absolute;
-  top: 100%;
-  right: 2%;
-  z-index: 1;
-  padding: 1rem;
-  background-color: #fff;
-  font-size: 1rem;
-  font-weight: 700;
-  border: 1px solid gray;
-  white-space: nowrap;
-`;
-const MenuLink = styled(Link)`
-  text-decoration: none;
-  color: black;
-`;
-const MenuLinkContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-`;
-const SignoutButton = styled.div`
-  &:hover {
-    cursor: pointer;
   }
 `;
